Flytter kode fra LastOppVedlegg2 til ny useFilopplaster2.tsx
