+++ src/frontend/components/SøknadsSteg/Dokumentasjon/LastOppVedlegg2.tsx
@@ -1,24 +1,11 @@
-import React, { useState } from 'react';
+import React from 'react';
 
-import {
-    Checkbox,
-    Detail,
-    FileAccepted,
-    FileObject,
-    FileRejected,
-    FileRejectionReason,
-    FileUpload,
-    FormSummary,
-    Heading,
-    List,
-    VStack,
-} from '@navikt/ds-react';
+import { Checkbox, FileUpload, FormSummary, Heading, List, VStack } from '@navikt/ds-react';
 
 import { useApp } from '../../../context/AppContext';
 import {
     dokumentasjonsbehovTilBeskrivelseSanityApiNavn,
     dokumentasjonsbehovTilTittelSanityApiNavn,
-    EFiltyper,
     IDokumentasjon,
     IVedlegg,
 } from '../../../typer/dokumentasjon';
@@ -26,14 +13,7 @@ import { Dokumentasjonsbehov } from '../../../typer/kontrakt/dokumentasjon';
 import { LocaleRecordBlock, Typografi } from '../../../typer/sanity/sanity';
 import { slåSammen } from '../../../utils/slåSammen';
 import TekstBlock from '../../Felleskomponenter/Sanity/TekstBlock';
-import { useLastRessurserContext } from '../../../context/LastRessurserContext';
-import axios from 'axios';
-import Miljø from '../../../../shared-utils/Miljø';
-
-interface OpplastetVedlegg {
-    dokumentId: string;
-    filnavn: string;
-}
+import { useFilopplaster2 } from './filopplaster/useFilopplaster2';
 
 interface Props {
     dokumentasjon: IDokumentasjon;
@@ -45,8 +25,20 @@ interface Props {
 }
 
 const LastOppVedlegg2: React.FC<Props> = ({ dokumentasjon, oppdaterDokumentasjon }) => {
+    const {
+        avvisteFiler,
+        MAKS_FILSTØRRELSE_MB,
+        MAKS_FILSTØRRELSE_BYTES,
+        MAKS_ANTALL_FILER,
+        støttedeFiltyper,
+        feilMeldinger,
+        leggTilVedlegg,
+        fjernVedlegg,
+        fjernAvvistFil,
+        fjernAlleAvvisteFiler,
+    } = useFilopplaster2(dokumentasjon, oppdaterDokumentasjon);
+
     const { søknad, tekster, plainTekst } = useApp();
-    const { wrapMedSystemetLaster } = useLastRessurserContext();
 
     const dokumentasjonstekster = tekster().DOKUMENTASJON;
     const frittståendeOrdTekster = tekster().FELLES.frittståendeOrd;
@@ -55,6 +47,9 @@ const LastOppVedlegg2: React.FC<Props> = ({ dokumentasjon, oppdaterDokumentasjon
         const huketAv = event.target.checked;
         const vedlegg = huketAv ? [] : dokumentasjon.opplastedeVedlegg;
         oppdaterDokumentasjon(dokumentasjon.dokumentasjonsbehov, vedlegg, huketAv);
+        if (huketAv) {
+            fjernAlleAvvisteFiler();
+        }
     };
 
     const tittel: LocaleRecordBlock =
@@ -70,88 +65,6 @@ const LastOppVedlegg2: React.FC<Props> = ({ dokumentasjon, oppdaterDokumentasjon
     const dokumentasjonsbeskrivelse = dokumentasjonsbehovTilBeskrivelseSanityApiNavn(
         dokumentasjon.dokumentasjonsbehov
     );
-
-    const [avvisteFiler, setAvvsiteFiler] = useState<FileRejected[]>([]);
-
-    const MAKS_FILSTØRRELSE_MB = 10;
-    const MAKS_FILSTØRRELSE_BYTES = MAKS_FILSTØRRELSE_MB * 1024 * 1024;
-    const MAKS_ANTALL_FILER = 25;
-
-    const støttedeFiltyper = [EFiltyper.PNG, EFiltyper.JPG, EFiltyper.JPEG, EFiltyper.PDF];
-
-    const feilMeldinger: Record<FileRejectionReason, string> = {
-        fileType: 'Filformatet støttes ikke',
-        fileSize: `Filen er større enn ${MAKS_FILSTØRRELSE_MB} MB`,
-    };
-
-    const dagensDatoStreng = new Date().toISOString();
-
-    const leggTilVedlegg = async (nyeFiler: FileObject[]) => {
-        const nyeAksepterteFiler = nyeFiler.filter(file => !file.error);
-        const nyeAvvisteFiler = nyeFiler.filter(file => file.error);
-
-        setAvvsiteFiler([...avvisteFiler, ...nyeAvvisteFiler]);
-
-        const nyeVedlegg: IVedlegg[] = [];
-
-        await Promise.all(
-            nyeAksepterteFiler.map((fil: FileAccepted) => {
-                return wrapMedSystemetLaster(async () => {
-                    const requestData = new FormData();
-                    requestData.append('file', fil.file);
-
-                    await axios
-                        .post<OpplastetVedlegg>(
-                            `${Miljø().dokumentProxyUrl}/mapper/familievedlegg`,
-                            requestData,
-                            {
-                                withCredentials: true,
-                                headers: {
-                                    'content-type': 'multipart/form-data',
-                                    accept: 'application/json',
-                                },
-                            }
-                        )
-                        .then((response: { data: OpplastetVedlegg }) => {
-                            const { data } = response;
-                            nyeVedlegg.push({
-                                dokumentId: data.dokumentId,
-                                navn: fil.file.name,
-                                størrelse: fil.file.size,
-                                tidspunkt: dagensDatoStreng,
-                                fil: fil.file,
-                            });
-                        });
-                });
-            })
-        );
-
-        if (nyeVedlegg.length > 0) {
-            oppdaterDokumentasjon(
-                dokumentasjon.dokumentasjonsbehov,
-                [...dokumentasjon.opplastedeVedlegg, ...nyeVedlegg],
-                dokumentasjon.harSendtInn
-            );
-        }
-    };
-
-    const fjernVedlegg = (vedleggSomSkalFjernes: IVedlegg) => {
-        const nyVedleggsliste = dokumentasjon.opplastedeVedlegg.filter(
-            (opplastetVedlegg: IVedlegg) => {
-                return opplastetVedlegg.dokumentId !== vedleggSomSkalFjernes.dokumentId;
-            }
-        );
-
-        oppdaterDokumentasjon(
-            dokumentasjon.dokumentasjonsbehov,
-            nyVedleggsliste,
-            dokumentasjon.harSendtInn
-        );
-    };
-
-    const fjernAvvistFil = (fil: FileRejected) => {
-        setAvvsiteFiler(avvisteFiler.filter(file => file !== fil));
-    };
 
     return (
         <FormSummary>
